Forked SimOS to SimFork
I don't really like SimOS anymore. It's mainly because of the convoluted
way I made apt and the registry work, but maybe it's just that a simple
menu isn't visually appealing nor flexible.

So, I forked SimOS and started deconstructing it. My new goal is to make
a shell rooted in UNIX.
I'm not quite sure what deatures will be in there. Permissions would be
a tough one to tackle.
This doesn't mean I'm killing SimOS, just that I want a break from it.
Maybe I'll revisit the idea when SimFork gets underway a bit.
SimFork is a working title, so I'll probably change it once it starts
developing.

Also, I fixed a very obvious security exploit in JMailServer. The only
problem is that if it hangs, you can't terminate it.