Added the process API
The process API has 4 functions: create, info, close, and run.
Create creates a process and adds it to the process list (or just runs
it syncronously, if you say so.)
info gets info on a process by PID or coroutine.
Close either sends a terminate event to a process or destoys it
immediately, depending on what you specify.
Run is an internal use function that start process' management of, well,
processes. It is called by init and will be ignored if called again.

Oh yeah, and /var/log is now being ignored.