Made a SimFork shell... sort of
A better phrase would be "lacking clone of the CraftOS shell". It's more
of a mockup than anything. I'll add unique features eventually, but for
now it's just crap.

As for crap, I went through several revisions, all of which either froze
or crashed the ingame computer due to mistakes in the process API and my
repeated accidental overwriting of the shell API. Oh well.