Added a login thingy
It's purely cosmetic for now. You can't add new users or change
passwords, and there is no file security. At leat it hashes the
passwords a bit.

You may notice that there is code to prevent termination, but that it's
incomplete. The reason being that it caused the read function to
shutdown the ingame computer for some reason. I'll look into it later.

I've also removed SimOS from the boot menu due to issues with the system
API namespace being reused improperly.