Added partial fs permissions to SimFork
Right now, only fs.list is implemented. But now that I've got the basic
framework down, it's just a bit of copy pasting. Except for fs.open, I
might have to do some bit checking for that one.

Fsperm details all the filesystem permissions, and mkperms recreates
that file with all owners set to root and all permissions set to 322.
Permissions are only 2 bits since reading and execution are basically
the same thing in Lua.

I also fixed some bugs with various parts of the shell.