This again
Another attempt to bring the convoluted and scattered bits of code
together in a single unreadable but easy to execute file. I did this
because I felt I was focusing too much on the UI, but I neede to figure
out the basics first.

This might fall through like the others. It's quite laughable in it's
current state as a mere copy-pate job, a reboot only in name. But it's
worth a shot, since I don't just want to dispose of what little I've
already done.