Updated JMail
Updated JMailServer to use a config file and JMail(Client) to scan for
servers by hostname even though this isn't implemented in the server
yet.
Please note that the config file will lose its comments if saved from
within the program.
On a side note, I've noticed several places in the 1.12.2 beta where an
"l" suffix is taped on variables where it shouldn't be, leading to
errors. I've noticed it in rednet.receive and io.lines so far.