Removed SimOS for good
I've decided that I'm done with it, so I threw it in an archive folder
that contains most of my other old crap. If you want something out of
that folder, ask. Anyway, I also added a trigger sequence for a debug
console in the process API, although I haven't added the actual console.
I'd probably have it accept input from single keys, much like the magic
SysRq key. Except that this won't interfere with normal system
operation. I also added a DH key exchange framework, but it's unusable
due to the shortcomings of floating point numbers and the modulo
operator.