Improved secses a little bit
It's now optimized a bit more. For instance, it uses a faster
equation to check for primitive roots and now only checks prime
primitive roots (which is what we want), but it still falls victim
to that old problem with floats. Why Lua, why? Why must every num
be a float? Can't we have some normal longs or shorts or ints?