Rewrote sendmail functions
JMail now has a proper headers system. There are a few problems with
this, however. The first is that viewing any old mail from before this
system will crash. The second is that registering an account is buggy
because sendmail can't parse the old system. (It doesn't crash sendmail,
but it does crash the client. The account is still registered.) The
third is that the reply function doesn't work (I have yet to test reply
all, but that probably fails too.) So don't upgrade yet.

I also fixed a bug in time that would cause a crash on 1.8 versions of
Minecraft.

Please don't try to read the code with gitlog. Due to a formatting error
caused by a stuck control key, I had to realign almost every code block,
which has the side effect of making the gitlog absolutely unreadble.