Fixed some bugs with cross-server communication
It now works AS LONG AS the user exists on the other end. If an error
like that happens, the two server will engage in a 'shouting match',
where they will keep pinging the same message about an incorrectly
formatted email. This was because I wrote the error handler before I
finalized the sendmail code, so it uses an outdated protocol.

Things I plan to do:
- Make date calculation server side,
- Fix errors with sendmail to eliminate shouting matches
- Reenable the terminate event when the console exits
- Fix occasional hanging shutdown
- Fix occasional registering client crash
- Allow client to cancel sending a mail
- Add per user disk limits (?)