Touched up SimFork shell
It now supports quoting of arguments and the && symbol. This also adds a
mofied version of shell.run(), and fixes a long-standing error with
secfs.

Not that it matters much. I don't think I'm going to continue
development on SimFork. Maybe I'll make a shell and APT, but that's it.
Implementing systemd with my existing code would not work too well, and
seeing as this project already failed once, I think it should get a real
shelving instead of the previous rebranding.

Perhaps, when I complete APT (if ever), I'll bring back SimOS. But who
knows?